Refactor classes to separate module
